Revert "[FIX]: InputOutput 파일 분리"
This reverts commit b49fc62c99ee440685701d9e59d8702c7f56370d.
